Update 2013-10-14-introducing-cfpb-open-tech.md
QUESTION: Where did the quote in paragraph two come from? Should source.

COMMENTS: some minor streamlining/first-personing of the voice; focused on "design" led by the Design Team, rather than the Design Team's work (we do process and policy design that we also want to release eventually, but which currently doesn't start with that team); reordered a few sentences/paragraphs for flow; removed "for ourselves" from the end, even though I get it because I think we should always be able to say that better operations for us means better service for the public.